% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{bump_version}
\alias{bump_version}
\title{Bump a version string semantically}
\usage{
bump_version(version, bump)
}
\arguments{
\item{version}{Version string, e.g. \code{"1.2.3"}.}

\item{bump}{One of \code{"major"}, \code{"minor"}, \code{"patch"}.}
}
\description{
Bump a version string semantically
}
